Classes that employ lazy evalution should have no friends. I removed all friends from the Jet class and then rewrote the source code files of the ex-friends to compensate. Unfortunately, this required changing the MXYZPTLK header files: new functionality, new inline functions, and so forth.  I regret that any module that uses MXYZPTLK will have to be rebuilt.
Note the new functions:

  Jet::resetIterator, Jet::resetConstIterator,
  Jet::stepIterator, and Jet::stepConstIterator

(and similarly for JetC). They may be useful. I'm thinking of
introducing the same functions in the beamline class to eliminate the
need for instantiating dlist_iterators in application programs.

However, be aware of the danger: the iterator is not robust against
changes in the Jet (or JetC). If you are using Jet::stepIterator, for
example, and modify the Jet between consecutive steps, almost anything
can happen.

- Leo Michelotti