I have extended to rbend the possibility of using a (telescoped) polynomial in place of the asin function. Because the rbend uses several propagators, this had to be done a little differently from what was done for sbend. In order to activate the polynomial approximation, add a line
  rbend::NoEdge.makeApproximate();

to your code. To undo this and return to the "Exact" propagator:

  rbend::NoEdge.makeExact();

There is a query function that indicates whether or not the polynomial
approximation is being used.

  if( rbend::NoEdge.isApproximate() ) cout << "yes" << endl;
  else                                cout << "no"  << endl;

This should make the polynomial approximation available for CF_rbends
as well. If you want to use the approximation in some rbends and not
in others, then you have to instantiate another functor to replace
rbend::NoEdge.

Leo Michelotti