---------------------------------------------------------------------- Committing in fnal/beamline.
Added Files:
  include/bmlnElmnt.h include/drift.h

Modified Files:
  include/CF_rbend.h include/Slot.h include/beamline.h
  include/beamline.inc include/combinedFunction.h include/kick.h
  include/mwiremonitor.h include/pinger.h
  include/rastermonitor.h include/tclmonitor.h

  src/BBLens.cc
  src/CF_rbend.cc src/JetQuadrupole.cc src/JetThinQuad.cc
  src/JetthinSext.cc src/Makefile src/Slot.cc src/beamline.cc
  src/bmlnElmnt.cc src/circuit.cc src/combinedFunction.cc
  src/drift.cc src/hkick.cc src/hmonitor.cc src/kick.cc
  src/marker.cc src/monitor.cc src/mwiremonitor.cc
  src/octupole.cc src/pinger.cc src/quadrupole.cc
  src/rastermonitor.cc src/rbend.cc src/sbend.cc src/sector.cc
  src/sextupole.cc src/srot.cc src/tclmonitor.cc
  src/thin12pole.cc src/thin14pole.cc src/thin16pole.cc
  src/thin18pole.cc src/thinDecapole.cc src/thinLamb.cc
  src/thinMultipole.cc src/thinOctupole.cc src/thinQuad.cc
  src/thinSeptum.cc src/thinSextupole.cc src/thinrfcavity.cc
  src/vkick.cc src/vmonitor.cc

  src/physics/6d/bmlnElmntPhysics.cc
----------------------------------------------------------------------

This is (only) the first step in breaking up the beamline.h
file. There is much that remains to be done, but I am trying to do it
all in such a way that everything continues working at each stage of
the development. At the moment, everything at least compiles under
Solaris in my directories. I hope it will in others.

To be "safe," I tagged all files in the repository "before_breakup"
before making these changes. Thus, we should be able to recover
in case of a catastrophe.

Note: I have also modified the Name() and Type() functions so that
they return a const char* rather than a char*.

Merry Christmas,
Leo Michelotti