Changes include modifications in bmlfactory.cc (thick monitors, combined function bends, etc.) which I've unsuccessfully tried to commit several times before.
Changes made to files XXX_table.c, viz.,

    g_hash_table_foreach -> g_hash_table_foreach_remove

reflect an attempt to fix a segmentation fault. Making this change in
function XXX_table_delete does not fix the fault; making it in
XXX_table_to_array does. This makes no sense to me. The problem is not
fixed, but the error has disappeared for now.

In the process of studying the bug, diagnostic lines were added
and then commented out. To find them, search for "DGN".

Leo Michelotti