Removing exit statements; replacing them with exceptions.  Within the C files, I've punted by introducing a C++ function called
  extern "C" void bmlfactory_exit();

and substituted it for all calls to exit(...).

  That having been done, there was one final wrinkle. The file mad.c
is written using flex on mad_lex.l:

  flex -B -8 -omad.c mad_lex.l

I modified mad_lex.l - and mad_yacc.y - so that "bmlfactory_exit();"
is written wherever "exit( EXIT_FAILURE );" used to appear. (For the
record: I did the same for the files mad.lex and mad.y although they
are no longer used. Shall we get rid of them?)

  However, that was not sufficient. In implementing a function called
yy_fatal_error, flex itself automatically creates an exit statement in
mad.c:

  #ifdef YY_USE_PROTOS
  static void yy_fatal_error( yyconst char msg[] )
  #else
  static void yy_fatal_error( msg )
  char msg[];
  #endif
  {
    (void) fprintf( stderr, "%s\n", msg );
    exit( YY_EXIT_FAILURE );
  }

To fix this, after mad.c is written, it can be processed as follows:

  echo "extern void bmlfactory_exit();" > mad.xxx
  sed -e "s/exit( YY_EXIT_FAILURE )/bmlfactory_exit()/" mad.c >> mad.xxx
  mv mad.xxx mad.c

This should be added to the build procedure so that it happens
automatically.

Leo Michelotti (aka lpjm)