I removed the lines marked "// REMOVE:."
According to the rules, that sentence is punctuated correctly
but leaves the impression that the period is part of the mark.
If one were, I suppose it would have to be punctuated "// REMOVE:.."

Or should I have written "// REMOVE:..." in the previous sentence?

And '"// REMOVE:..."' in the previous one?

And "'"// REMOVE:..."'" in the previous one?

  ...

- lpjm