Migrating changes already made in branch StRel20070808-patches on January 10.
For the purpose of this note, a "regular bend" is one whose faces are
either normal entry or parallel; otherwise, I'll call it "irregular."
Before January 10, DriftsToSlots worked only with regular bends.  I am
generalizing it to work with irregular bends as well.

** WARNING: ONLY PARTLY DONE!! ***  This version still fails (a) for
non-symmetric irregular bends: i.e. those that have DIFFERENT entry
and exit edge angles.  (b) possibly, if an irregular bend is rolled.
(c) when a drift is both upstream and downstream of an irregular bend.

- lpjm