Removed the "KLUDGE" that prevented .setup from working more than once...
Removed the "KLUDGE" that prevented the XXX::setup routine from working
more than once on the same or a cloned CF_rbend. To do this, an
attribute, "integrated_strengths_," was added to capture integrated
values of multipole components.  (It replaced "multipoles_" -- which
had never been used anyway.)

As a reminder: the issue of multiply redundant setups has not been resolved.