Wrote missing "non-const" implementation for beamline::OrbitLength(...)
A const version of bmlElmnt::OrbitLength(...) was added to the base class
header file bmlnElmnt.h in April, 2014. (See log for commit 5277ecbbde...)
  ---------------------------
  -> git diff 44619d41e 5277ec beamline/include/bmlnElmnt.h
  ...
  +  virtual double OrbitLength( Particle const& ) const { return length_; }
     virtual double OrbitLength( Particle const& ) { return length_; }
  ---------------------------
However, only the const version was implemented in the source file
beamline/src/beamline.cc.
  ---------------------------
  -> egrep -i 'beamline\ *::\ *orbitlength' beamline/src/beamline.cc
  double beamline::OrbitLength( Particle const& x ) const
  ---------------------------
In an application, the compiler could and did default to the code in
the base class header file for the non-const implementation.